refactor: improve the folders structure
