+++ src/components/product/OrderNowBtn.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { FaArrowRight } from 'react-icons/fa6';
-import { Button } from './ui/button';
+import { Button } from '@/components/ui/button';
 
 interface OrderNowBtnProps {
   onClick: () => void;
