+++ src/components/product/ProductCard.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import React, { useState } from 'react';
-import { OrderNowBtn } from '@/components/OrderNowBtn';
-import { ColorDots } from '@/components/ColorDots';
-import { StockStatus } from '@/components/StockStatus';
+import { OrderNowBtn } from '@/components/product/OrderNowBtn';
+import { ColorDots } from '@/components/product/ColorDots';
+import { StockStatus } from '@/components/product/StockStatus';
 import Image from 'next/image';
 
 export const ProductCard: React.FC<ProductCardProps> = ({
@@ -48,7 +48,7 @@ export const ProductCard: React.FC<ProductCardProps> = ({
             />
           </figcaption>
         </figure>
-        <p className="min-h-[72px]">{shortDescription}</p>
+        <p className='min-h-[72px]'>{shortDescription}</p>
         <div className='flex items-center justify-between'>
           <p className='font-base font-bold text-grey-900'>
             {pricePerMonth}€/month
