+++ src/components/product/ProductGrid.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ProductCard } from '@/components/ProductCard';
+import { ProductCard } from '@/components/product/ProductCard';
 import { products } from '@/data/mockData';
 
 export const ProductGrid: React.FC = () => {
