+++ src/components/product/StockStatus.tsx
